moved parseResult from teraslice-utils class to display
