+++ packages/teraslice-cli/test/cmds/ex/errors-spec.js
@@ -19,6 +19,7 @@ describe('errors', () => {
             expect(yargsResult.clusterAlias).toEqual('ts-test1');
             expect(yargsResult.id).toEqual('99999999-9999-9999-9999-999999999999');
         });
+
         it('should parse properly with args', () => {
             const yargsCmd = yargs.command(
                 errors.command,
